// examples/experiments/a2ui-private/original_a2ui_source/renderers/lit/src/0.8/types/types.ts
export interface TextNode extends BaseComponentNode {
  type: "Text";
  properties: ResolvedText;
}

export interface ImageNode extends BaseComponentNode {
  type: "Image";
  properties: ResolvedImage;
}

export interface IconNode extends BaseComponentNode {
  type: "Icon";
  properties: ResolvedIcon;
}

export interface VideoNode extends BaseComponentNode {
  type: "Video";
  properties: ResolvedVideo;
}

export interface AudioPlayerNode extends BaseComponentNode {
  type: "AudioPlayer";
  properties: ResolvedAudioPlayer;
}

export interface RowNode extends BaseComponentNode {
  type: "Row";
  properties: ResolvedRow;
}

export interface ColumnNode extends BaseComponentNode {
  type: "Column";
  properties: ResolvedColumn;
}

export interface ListNode extends BaseComponentNode {
  type: "List";
  properties: ResolvedList;
}

export interface CardNode extends BaseComponentNode {
  type: "Card";
  properties: ResolvedCard;
}

export interface TabsNode extends BaseComponentNode {
  type: "Tabs";
  properties: ResolvedTabs;
}

export interface DividerNode extends BaseComponentNode {
  type: "Divider";
  properties: ResolvedDivider;
}

export interface ModalNode extends BaseComponentNode {
  type: "Modal";
  properties: ResolvedModal;
}

export interface ButtonNode extends BaseComponentNode {
  type: "Button";
  properties: ResolvedButton;
}

export interface CheckboxNode extends BaseComponentNode {
  type: "CheckBox";
  properties: ResolvedCheckbox;
}

export interface TextFieldNode extends BaseComponentNode {
  type: "TextField";
  properties: ResolvedTextField;
}

export interface DateTimeInputNode extends BaseComponentNode {
  type: "DateTimeInput";
  properties: ResolvedDateTimeInput;
}

export interface MultipleChoiceNode extends BaseComponentNode {
  type: "MultipleChoice";
  properties: ResolvedMultipleChoice;
}

export interface SliderNode extends BaseComponentNode {
  type: "Slider";
  properties: ResolvedSlider;
}

export interface CustomNode extends BaseComponentNode {
  type: string;
  // For custom nodes, properties are just a map of string keys to any resolved value.
  properties: CustomNodeProperties;
}

/**
 * The complete discriminated union of all possible resolved component nodes.
 * A renderer would use this type for any given node in the component tree.
 */
export type AnyComponentNode =
  | TextNode
  | IconNode
  | ImageNode
  | VideoNode
  | AudioPlayerNode
  | RowNode
  | ColumnNode
  | ListNode
  | CardNode
  | TabsNode
  | DividerNode
  | ModalNode
  | ButtonNode
  | CheckboxNode
  | TextFieldNode
  | DateTimeInputNode
  | MultipleChoiceNode
  | SliderNode
  | CustomNode;
// ...
export interface ResolvedColumn {
  children: AnyComponentNode[];
  distribution?:
    | "start"
    | "center"
    | "end"
    | "spaceBetween"
    | "spaceAround"
    | "spaceEvenly";
  alignment?: "start" | "center" | "end" | "stretch";
}